slm_finetune_framework/pyproject.toml
[project]
name = "slm_finetune_framework"
version = "0.1.0"
description = "Config-driven pipeline for finetuning small language models with LoRA."
authors = [{name = "AutoGen"}]
requires-python = ">=3.11"
dependencies = [
  "transformers",
  "peft",
  "torch",
  "sqlalchemy",
  "pyyaml",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
